refactor: qc-core subscribe 도메인 패키지 정리
