+++ qc-application/src/main/java/com/eager/questioncloud/application/api/subscribe/SubscribeProcessor.java
@@ -1,8 +1,8 @@
 package com.eager.questioncloud.application.api.subscribe;
 
 import com.eager.questioncloud.core.domain.creator.infrastructure.CreatorRepository;
-import com.eager.questioncloud.core.domain.subscribe.Subscribe;
-import com.eager.questioncloud.core.domain.subscribe.SubscribeRepository;
+import com.eager.questioncloud.core.domain.subscribe.infrastructure.SubscribeRepository;
+import com.eager.questioncloud.core.domain.subscribe.model.Subscribe;
 import com.eager.questioncloud.exception.CustomException;
 import com.eager.questioncloud.exception.Error;
 import lombok.RequiredArgsConstructor;
