+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/subscribe/event/UnsubscribedEvent.java
@@ -1,4 +1,4 @@
-package com.eager.questioncloud.core.domain.subscribe;
+package com.eager.questioncloud.core.domain.subscribe.event;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
