+++ qc-core/src/main/java/com/eager/questioncloud/core/domain/subscribe/infrastructure/SubscribeEntity.java
@@ -1,5 +1,6 @@
-package com.eager.questioncloud.core.domain.subscribe;
+package com.eager.questioncloud.core.domain.subscribe.infrastructure;
 
+import com.eager.questioncloud.core.domain.subscribe.model.Subscribe;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
